+++ projects/engine/nex/pbr/GlobalIllumination.cpp
@@ -494,7 +494,7 @@ nex::GlobalIllumination::~GlobalIllumination() = default;
 
 void nex::GlobalIllumination::bakeProbes(const Scene & scene, Renderer* renderer)
 {
-	const size_t size = 1024;
+	const int size = 1024;
 
 	PerspectiveCamera camera(1.0f, glm::radians(90.0f), 0.1f, 100.0f);
 	//OrthographicCamera camera(1.0f, 1.0f, 0.1f, 100.0f);
@@ -614,7 +614,7 @@ void nex::GlobalIllumination::bakeProbes(const Scene & scene, Renderer* renderer
 
 void nex::GlobalIllumination::bakeProbe(ProbeVob* probeVob, const Scene& scene, Renderer* renderer)
 {
-	const size_t size = 1024;
+	const int size = 1024;
 
 	PerspectiveCamera camera(1.0f, glm::radians(90.0f), 0.1f, 100.0f);
 	//OrthographicCamera camera(1.0f, 1.0f, 0.1f, 100.0f);
